+++ backend/ops_api/ops/__init__.py
@@ -5,11 +5,11 @@ from typing import Optional
 from flask import Blueprint
 from flask import Flask
 from flask_cors import CORS
-from ops.auth.utils import jwtMgr
-from ops.auth.utils import oauth
 from ops.home_page.views import home
-from ops.models.users import db
+from ops.models.base import db
 from ops.urls import register_api
+from ops.utils.auth import jwtMgr
+from ops.utils.auth import oauth
 
 
 def configure_logging() -> None:
@@ -45,7 +45,6 @@ def create_app(config_overrides: Optional[dict] = None) -> Flask:
     if config_overrides is not None:
         app.config.from_mapping(config_overrides)
 
-    # app.register_blueprint(ops.summary.urls.bp)
     app.register_blueprint(home)
 
     api_bp = Blueprint(
