+++ backend/ops_api/ops/models/cans.py
@@ -1,7 +1,7 @@
+from ops.models.base import BaseModel
+from ops.models.base import db
 from ops.models.portfolios import Portfolio
 from ops.models.portfolios import portfolio_cans
-from ops.utils import BaseModel
-from ops.utils import db
 from sqlalchemy.engine import Connection
 from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy.orm import column_property
