+++ backend/ops_api/tests/ops/auth/test_oidc_controller.py
@@ -2,7 +2,7 @@ from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.hazmat.primitives.serialization import Encoding
 from cryptography.hazmat.primitives.serialization import NoEncryption
 from cryptography.hazmat.primitives.serialization import PrivateFormat
-from ops.auth.utils import get_jwt
+from ops.utils.auth import get_jwt
 import pytest
